+++ ErsatzTV.Application/Maintenance/Commands/DeleteOrphanedArtworkHandler.cs
@@ -3,16 +3,12 @@ using ErsatzTV.Core.Interfaces.Repositories;
 
 namespace ErsatzTV.Application.Maintenance;
 
-public class DeleteOrphanedArtworkHandler : IRequestHandler<DeleteOrphanedArtwork, Either<BaseError, Unit>>
+public class DeleteOrphanedArtworkHandler(IArtworkRepository artworkRepository)
+    : IRequestHandler<DeleteOrphanedArtwork, Either<BaseError, Unit>>
 {
-    private readonly IArtworkRepository _artworkRepository;
-
-    public DeleteOrphanedArtworkHandler(IArtworkRepository artworkRepository) =>
-        _artworkRepository = artworkRepository;
-
     public Task<Either<BaseError, Unit>>
         Handle(DeleteOrphanedArtwork request, CancellationToken cancellationToken) =>
-        _artworkRepository.GetOrphanedArtwork()
-            .Bind(_artworkRepository.Delete)
+        artworkRepository.GetOrphanedArtworkIds()
+            .Bind(artworkRepository.Delete)
             .Map(_ => Right<BaseError, Unit>(Unit.Default));
 }