+++ ErsatzTV.Core/Interfaces/Repositories/IArtworkRepository.cs
@@ -1,9 +1,7 @@
-﻿using ErsatzTV.Core.Domain;
-
-namespace ErsatzTV.Core.Interfaces.Repositories;
+﻿namespace ErsatzTV.Core.Interfaces.Repositories;
 
 public interface IArtworkRepository
 {
-    Task<List<Artwork>> GetOrphanedArtwork();
-    Task<Unit> Delete(List<Artwork> artwork);
+    Task<List<int>> GetOrphanedArtworkIds();
+    Task<Unit> Delete(List<int> artworkIds);
 }